+++ Repositories/UnitOfWork.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading.Tasks;
 using KamatekCrm.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Storage;
@@ -40,6 +41,11 @@ namespace KamatekCrm.Repositories
         public int SaveChanges()
         {
             return _context.SaveChanges();
+        }
+
+        public async Task<int> SaveChangesAsync()
+        {
+            return await _context.SaveChangesAsync();
         }
 
         /// <summary>